requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
pyQiwiP2P==2.0.6